refactor: refactored LoginScreen component by dividing it in multiple components
