+++ source/navigator/AuthStack.tsx
@@ -1,7 +1,7 @@
 import React, { useEffect, useState } from "react";
 import { createNativeStackNavigator } from "@react-navigation/native-stack";
+import LoginScreen from "../screens/loginScreen/LoginScreen";
 import Onboarding from "../screens/onboarding";
-import LoginScreen from "../screens/LoginScreen";
 import StorageService, {
   ONBOARDING_DISABLED,
 } from "../services/StorageService";
