+++ source/screens/index.js
@@ -1,6 +1,6 @@
 export { default as DevFeaturesScreen } from "./DevFeaturesScreen";
 export { default as FormCaseScreen } from "./FormCaseScreen";
-export { default as LoginScreen } from "./LoginScreen";
+export { default as LoginScreen } from "./loginScreen/LoginScreen";
 export { default as ProfileScreen } from "./profileScreen/ProfileScreen";
 export { default as SplashScreen } from "./SplashScreen";
 export { default as AboutScreen } from "./AboutScreen";
