+++ source/screens/loginScreen/LoginScreen.types.ts
@@ -1,0 +1,5 @@
+import type { ThemeType } from "../../styles/themeHelpers";
+
+export interface ApiStatusMessageContainerProps {
+  theme: ThemeType;
+}
